% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_process_f.R
\name{data_process_f}
\alias{data_process_f}
\title{This function converting OS time (in days) into months and removing samples where OS time information is missing. Note: In the Example data OS time labelled is "OS.time" and OS event labelled as "OS"}
\usage{
data_process_f(data, col_num, output)
}
\arguments{
\item{output}{: arg3 - name of output file, in chich user want to store data}

\item{data:args1}{\itemize{
\item data (Patients data with clinical and gene expression, where samples are in rows and features/genes are in columns)
}}

\item{col_num:args2}{\itemize{
\item column number in data at where clinical info ends
}}
}
\description{
This function converting OS time (in days) into months and removing samples where OS time information is missing. Note: In the Example data OS time labelled is "OS.time" and OS event labelled as "OS"
}
\examples{
SurvPredPipe::data_process_f(data="TCGA-LGG_protein_coding_FPKM_data_with_clin_data.txt",col_num=20, output="New_data.txt") 
Usage: data_process_f(data,  col_num, output)
}
