% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_median_surv_barplot_f.R
\name{mean_median_surv_barplot_f}
\alias{mean_median_surv_barplot_f}
\title{This function generate barplots for mean and median survival of patients. Besides, user can also highlight one specific sample by providing sample IDs}
\usage{
mean_median_surv_barplot_f(surv_mean_med_data, selected_sample)
}
\arguments{
\item{surv_mean_med_data}{:args1 - Predicted Mean median survival time data for patients}

\item{selected_sample}{:args2 - ID of the sample for which user want to highlight in the plot}
}
\description{
This function generate barplots for mean and median survival of patients. Besides, user can also highlight one specific sample by providing sample IDs
}
\examples{
mean_median_surv_barplot_f(surv_mean_med_data="../inst/extdata/mean_median_survival_time_data.txt", selected_sample="TCGA-TQ-A8XE-01")
usgae: mean_median_surv_barplot_f(surv_mean_med_data, selected_sample)
}
