% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_curve_plots_f.R
\name{surv_curve_plots_f}
\alias{surv_curve_plots_f}
\title{This function generate individual survival curve for patients. Moreover, user can also highlight one specific sample by providing sample IDs}
\usage{
surv_curve_plots_f(Surv_curve_data, selected_sample)
}
\arguments{
\item{Surv_curve_data}{:args1 - Survival probabilty data for survival curve of patients}

\item{selected_sample:args2}{\itemize{
\item ID of a specific sample user wants to highlights on the plot against the background of curves of other samples
}}
}
\description{
This function generate individual survival curve for patients. Moreover, user can also highlight one specific sample by providing sample IDs
}
\examples{
SurvPredPipe::surv_curve_plots_f(Surv_curve_data="survCurves_data.txt", selected_sample="TCGA-DH-5140-01")
usgae: surv_curve_plots_f(Surv_curve_data, selected_sample)
}
