% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tr_test_f.R
\name{tr_test_f}
\alias{tr_test_f}
\title{This function split data into training and test data as per user defined ratio}
\usage{
tr_test_f(data, fraction, train_data, test_data)
}
\arguments{
\item{fraction, }{by which user want to split data for training data; e.g. 90\% training, so fraction would be 0.9}

\item{train_data, }{name of training set that user can provide}

\item{test_data, }{name of test set that user can provide}

\item{data:}{a dataframe where features in the columns and samples must be in rows}
}
\value{
training and test data
}
\description{
This function split data into training and test data as per user defined ratio
}
\examples{
tr_test_f(data="../inst/extdata/New_data.txt",fraction=0.9, train_data="train_FPKM.txt",test_data="test_FPKM.txt")
Usage: tr_test_f(data, fraction, train_data, test_data)
}
