% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_test_normalization_f.R
\name{train_test_normalization_f}
\alias{train_test_normalization_f}
\title{This function first tranform FPKM data into log2-scale transformation, followed by quantile normalization, where training data values used as target matrix both for training and test data for quantile  normalization}
\usage{
train_test_normalization_f(
  train_data,
  test_data,
  col_num,
  train_clin_data,
  test_clin_data,
  train_Normalized_data_clin_data,
  test_Normalized_data_clin_data
)
}
\arguments{
\item{train_data}{:args1 - training data (TSV) (Patients data with clinical and gene expression, where samples are in rows and features/genes are in columns)}

\item{col_num}{:args3 - column number in data at where clinical info ends}

\item{test_data_data}{:args2 - test data (TSV) (Patients data with clinical and gene expression, where samples are in rows and features/genes are in columns)}

\item{train_clin_data:}{args4- name of training data output which stores only clinical information}

\item{test_clin_data:}{args5 - name of test data output which stores only clinical information}

\item{train_Normalized_data_clin_data:}{args6 -  output file name - outfile file containing training clinical data and normalized data  (which is log-scaled followed by quantile normalized data).}

\item{test_Normalized_data_clin_data:}{args7- output filename - outfile file containing test clinical data and normalized data  (which is log-scaled followed by quantile normalized data).}
}
\description{
This function first tranform FPKM data into log2-scale transformation, followed by quantile normalization, where training data values used as target matrix both for training and test data for quantile  normalization
}
\examples{
SurvPredPipe::train_test_normalization_f(train_data="train_FPKM.txt",test_data="test_FPKM.txt", col_num=21, train_clin_data="Train_Clin.txt", test_clin_data="TestClin.txt", train_Normalized_data_clin_data="Train_Norm_data.txt", test_Normalized_data_clin_data="Test_Norm_data.txt")
Usage: train_test_normalization_f(train_data,  test_data, col_num, train_clin_data, test_clin_data, train_Normalized_data_clin_data, test_Normalized_data_clin_data)
}
