% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Univariate_sig_features_f.R
\name{Univariate_sig_features_f}
\alias{Univariate_sig_features_f}
\title{This function selects significant features (p-value <0.05)s based on median expression values of features using Univariate Survival analysis.}
\usage{
Univariate_sig_features_f(
  train_data,
  test_data,
  col_num,
  surv_time,
  surv_event,
  output_univariate_train,
  output_univariate_test
)
}
\arguments{
\item{train_data}{:args1 - training data (Patients data with clinical and gene expression, where samples are in rows and features/genes are in columns)}

\item{col_num}{:args3 - column number in data at where clinical info ends}

\item{surv_time}{:arg4 - name of column which contain survival time (in days) information}

\item{surv_event}{:arg5 - name of column which contain survival eventinformation}

\item{output_univariate_train}{:args6- name of output to store univariate selected features for training data}

\item{output_univariate_test}{:args7- name of output to store univariate selected features for test data}

\item{test_data_data}{:args2 - training data (Patients data with clinical and gene expression, where samples are in rows and features/genes are in columns)}
}
\description{
This function selects significant features (p-value <0.05)s based on median expression values of features using Univariate Survival analysis.
}
\examples{
Univariate_sig_features_f(train_data="../inst/extdata/Train_Norm_data.txt", test_data="../inst/extdata/Test_Norm_data.txt", col_num=21, surv_time="OS_month" , surv_event="OS" ,output_univariate_train="Train_Uni_sig_data.txt", output_univariate_test="Test_Uni_sig_data.txt") 
Usage: Univariate_sig_features_f(train_data, test_data, col_num, surv_time, surv_event, output_univariate_train, output_univariate_test)
}
